+++ src/components/stable/gux-notification-toast/tests/gux-notification-toast.spec.ts
@@ -70,7 +70,7 @@ describe('gux-notification-toast', () => {
       const page = await newSpecPage({ components, html, language });
       const element = page.root as HTMLElement;
       const dismissButton = element.shadowRoot.querySelector(
-        'gux-dismiss-button-beta'
+        'gux-dismiss-button'
       ) as HTMLElement;
       const guxdismissSpy = jest.fn();
       const clickSpy = jest.fn();
@@ -98,7 +98,7 @@ describe('gux-notification-toast', () => {
       const page = await newSpecPage({ components, html, language });
       const element = page.root as HTMLElement;
       const dismissButton = element.shadowRoot.querySelector(
-        'gux-dismiss-button-beta'
+        'gux-dismiss-button'
       ) as HTMLElement;
       const elementRemoveSpy = jest.spyOn(element, 'remove');