+++ src/components/stable/gux-simple-toast/tests/gux-simple-toast.spec.ts
@@ -65,7 +65,7 @@ describe('gux-simple-toast', () => {
       const page = await newSpecPage({ components, html, language });
       const element = page.root as HTMLElement;
       const dismissButton = element.shadowRoot.querySelector(
-        'gux-dismiss-button-beta'
+        'gux-dismiss-button'
       ) as HTMLElement;
       const guxdismissSpy = jest.fn();
       const clickSpy = jest.fn();
@@ -92,7 +92,7 @@ describe('gux-simple-toast', () => {
       const page = await newSpecPage({ components, html, language });
       const element = page.root as HTMLElement;
       const dismissButton = element.shadowRoot.querySelector(
-        'gux-dismiss-button-beta'
+        'gux-dismiss-button'
       ) as HTMLElement;
       const elementRemoveSpy = jest.spyOn(element, 'remove');